requirements.txt
helpers==0.2.0
protobuf==5.29.3
python-dotenv==1.0.1
config==0.5.1
Flask==3.1.0
services==0.1.1
google-genai
html2text